# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SkillQuest",
    "description": "Игровой бэкенд SkillQuest: диагностика навыков, игровые миссии, рекомендации курсов и стажировок, анализ резюме. Защищённые операции требуют Bearer-токен (см. security и docs/API.md).",
    "contact": {
      "name": "SkillQuest Team",
      "url": "https://github.com/BrainL4g/max-edu-2026"
    },
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/users/{user_id}/consent": {
      "get": {
        "tags": [
          "privacy"
        ],
        "summary": "Get Consent",
        "description": "Текущее состояние согласия пользователя на обработку ПД.",
        "operationId": "get_consent_users__user_id__consent_get",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentOut"
                }
              }
            }
          },
          "401": {
            "description": "Нет или невалидный Bearer-токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён: не свой ресурс или не та роль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "privacy"
        ],
        "summary": "Give Consent",
        "description": "Зафиксировать согласие на обработку ПД (152-ФЗ ст. 9).",
        "operationId": "give_consent_users__user_id__consent_post",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentOut"
                }
              }
            }
          },
          "401": {
            "description": "Нет или невалидный Bearer-токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён: не свой ресурс или не та роль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "privacy"
        ],
        "summary": "Withdraw Consent",
        "description": "Отозвать согласие на обработку ПД (152-ФЗ ст. 9, п. 6).",
        "operationId": "withdraw_consent_users__user_id__consent_delete",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentOut"
                }
              }
            }
          },
          "401": {
            "description": "Нет или невалидный Bearer-токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён: не свой ресурс или не та роль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/test-users/reset": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Reset Test Users",
        "description": "Сбросить учебные данные тест-студентов (id и токены сохраняются).\n\nОчищает попытки, Skill Map, резюме и анализ резюме, снимает целевую\nроль и возвращает профиль к состоянию по умолчанию — чтобы прогоны\nDATA-API были воспроизводимыми. Требует роль ``admin``.",
        "operationId": "reset_test_users_admin_test_users_reset_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetOut"
                }
              }
            }
          },
          "401": {
            "description": "Нет или невалидный Bearer-токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён: не свой ресурс или не та роль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/assessment/questions": {
      "get": {
        "tags": [
          "assessment"
        ],
        "summary": "Get Assessment Questions",
        "description": "Банк вопросов диагностики: текст и варианты ответов.\n\nПубличный без ``user_id``. С ``user_id`` требуется Bearer-токен\nвладельца (иначе 401/403), вопросы фильтруются по направлению целевой\nроли пользователя (``target_role.direction``, иначе ``user.direction``).",
        "operationId": "get_assessment_questions_assessment_questions_get",
        "security": [],
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssessmentQuestionOut"
                  },
                  "title": "Response Get Assessment Questions Assessment Questions Get"
                }
              }
            }
          },
          "401": {
            "description": "Нет или невалидный Bearer-токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён: не свой ресурс или не та роль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssessmentQuestionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "skill": {
            "type": "string",
            "title": "Skill"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Options"
          }
        },
        "type": "object",
        "required": [
          "id",
          "skill",
          "text",
          "options"
        ],
        "title": "AssessmentQuestionOut",
        "description": "Вопрос диагностики (ответ API, без внутренних оценок уровней).",
        "examples": [
          {
            "id": 1,
            "options": [
              "Почти не писал(а): знаю только print",
              "Писал(а) простые скрипты: циклы, условия, списки"
            ],
            "skill": "Python",
            "text": "Как вы работаете с Python? Оцените свой уровень."
          }
        ]
      },
...
      "ConsentOut": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "consent_given": {
            "type": "boolean",
            "title": "Consent Given"
          },
          "consent_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent At"
          },
          "policy_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Version"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "consent_given"
        ],
        "title": "ConsentOut",
        "description": "Состояние согласия пользователя на обработку ПД.",
        "examples": [
          {
            "consent_at": "2026-09-29T10:00:00",
            "consent_given": true,
            "policy_version": "1.0",
            "user_id": 1
          }
        ]
      },
...
      "ErrorOut": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorOut",
        "description": "Ошибка API: человекочитаемое сообщение в поле ``detail``."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResetOut": {
        "properties": {
          "reset": {
            "type": "boolean",
            "title": "Reset"
          },
          "students_affected": {
            "type": "integer",
            "title": "Students Affected"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "reset",
          "students_affected",
          "message"
        ],
        "title": "ResetOut",
        "description": "Результат сброса учебных данных тест-студентов.",
        "examples": [
          {
            "message": "Учебные данные тест-студентов сброшены",
            "reset": true,
            "students_affected": 1
          }
        ]
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}